+++ ovk/core/ConnectivityR.h
@@ -30,7 +30,7 @@ struct ovk_connectivity_r {
   MPI_Comm comm;
   int comm_size;
   int comm_rank;
-  size_t count;
+  long long count;
   const ovk_grid *grid;
   t_connectivity_r_edits edits;
   int *points[MAX_DIMS];
